+++ src/shader/mod.rs
@@ -1,9 +1,9 @@
 pub mod constants;
+pub mod pre_processor;
 pub mod program;
-pub mod read;
 pub mod service;
 
 pub use constants::*;
+pub use pre_processor::*;
 pub use program::*;
-pub use read::*;
 pub use service::*;
